Create Currency_Converter_Function.py
This is a function that converts a foreign currency to shekels or shekels to a foreign currency. See line 8 for comments.
I added three inputs and 2 outputs to test this function, they are outside the function and are not needed. The function will return a list, the first element being the value after conversion, and the second elements being a string indicating what currency the output is in.